refactor: 폴더 구조 개선, ranking과 관련한 파일 분리, components 폴더 분리
