+++ src/pages/games/SnackGame/game/components/ExpBarChart.tsx
