+++ src/pages/games/SnackGame/game/components/GameResult.tsx
