+++ src/pages/games/SnackGame/game/components/SnackGameHUD.tsx
