+++ src/pages/games/SnackGame/ranking/components/OtherRanking.tsx
