+++ src/pages/games/SnackGame/ranking/components/TopRanking.tsx
