+++ src/pages/games/SnackGame/ranking/components/UserRanking.tsx
